+++ src/index.html
@@ -8,7 +8,7 @@
       integrity="sha384-vp86vTRFVJgpjF9jiIGPEEqYqlDwgyBgEF109VFjmqGmIY/Y4HV4d3Gp2irVfcrp"
       crossorigin="anonymous"
     />
-    <title>Chatlingo</title>
+    <title>ChatLingo</title>
   </head>
 
   <body>
